pyproject.toml
[project]
name = "metronome"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mido>=1.3.3",
    "numpy>=2.4.2",
    "pyaudio>=0.2.14",
    "pyfluidsynth>=1.3.4",
    "pygame>=2.6.1",
    "python-rtmidi>=1.5.8",
    "ruff>=0.15.0",
]
